+++ tests/framework/validators/ExistValidatorTest.php
@@ -198,7 +198,7 @@ abstract class ExistValidatorTest extends DatabaseTestCase
     {
         $val = new ExistValidator([
            'targetClass' => OrderItem::className(),
-           'targetAttribute' => ['id' => 'COALESCE(order_id, 0)'],
+           'targetAttribute' => ['id' => 'COALESCE([[order_id]], 0)'],
        ]);
 
         $m = new Order(['id' => 1]);
@@ -235,7 +235,7 @@ abstract class ExistValidatorTest extends DatabaseTestCase
         $val->validateAttribute($m, 'id');
         $this->assertTrue($m->hasErrors('id'));
     }
-    
+
     public function testForceMaster()
     {
         $connection = $this->getConnectionWithInvalidSlave();